+++ modules/server/index.js
@@ -55,13 +55,12 @@ module.exports.readConfig = function (configIn) {
 module.exports.setup = function (cb) {
 	if (config.protocol === 'https') {
 		// https
-		serverEngine.setup(function (error) {
+		return serverEngine.setup(function (error) {
 			if (error) {
 				return cb(error);
 			}
 			controller.setup(cb);
 		});
-		return;
 	}
 	
 	log.info('server protocol: ' + config.protocol);